+++ static/js/nzbhydra.js
@@ -1122,7 +1122,8 @@ function StatsController($scope, stats, $state) {
     $scope.avgResponseTimes = stats.avgResponseTimes;
     $scope.avgIndexerSearchResultsShares = stats.avgIndexerSearchResultsShares;
     $scope.avgIndexerAccessSuccesses = stats.avgIndexerAccessSuccesses;
-
+    $scope.indexerDownloadShares = stats.indexerDownloadShares;
+    
 
     $scope.tabs = [
         {