+++ ui-src/js/config-fields-service.js
@@ -130,7 +130,7 @@ function ConfigFields() {
                 }
             }
         ]);
-        
+
         if (testtype == "newznab") {
             fieldset.push(
                 {
@@ -154,7 +154,34 @@ function ConfigFields() {
                         label: 'API hit reset time',
                         help: 'Local time at which the API hit counter is reset'
                     }
-                })
+                });
+            fieldset.push(
+                {
+                    key: 'username',
+                    type: 'horizontalInput',
+                    hideExpression: '!model.enabled',
+                    templateOptions: {
+                        type: 'text',
+                        required: false,
+                        label: 'Username',
+                        help: 'Only needed if indexer requires HTTP auth for API access (rare)'
+                    }
+                }
+            );
+            fieldset.push(
+                {
+                    key: 'password',
+                    type: 'horizontalInput',
+                    hideExpression: '!model.enabled || !model.username',
+                    templateOptions: {
+                        type: 'text',
+                        required: false,
+                        label: 'Password',
+                        help: 'Only needed if indexer requires HTTP auth for API access (rare)'
+                    }
+                }
+            )
+
         }
 
 
@@ -576,7 +603,6 @@ function ConfigFields() {
                                 help: "Not all indexers provide this information"
                             }
                         },
-
                         {
                             key: 'ignoreWords',
                             type: 'horizontalInput',
@@ -587,7 +613,6 @@ function ConfigFields() {
                                 help: "Results with any of these words in the title will be ignored"
                             }
                         },
-
                         {
                             key: 'requireWords',
                             type: 'horizontalInput',
@@ -598,7 +623,6 @@ function ConfigFields() {
                                 help: "Only results with at least of these words in the title will be displayed"
                             }
                         },
-
                         {
                             key: 'maxAge',
                             type: 'horizontalInput',
@@ -611,8 +635,6 @@ function ConfigFields() {
                                 }
                             }
                         },
-
-
                         {
                             key: 'generate_queries',
                             type: 'horizontalMultiselect',
@@ -634,6 +656,7 @@ function ConfigFields() {
                                 required: true
                             }
                         }
+
                     ]
                 },
                 {
@@ -647,7 +670,7 @@ function ConfigFields() {
                             type: 'horizontalSelect',
                             templateOptions: {
                                 type: 'select',
-                                label: 'Type',
+                                label: 'HTML parser',
                                 options: [
                                     {name: 'Default BS (slow)', value: 'html.parser'},
                                     {name: 'LXML (faster, needs to be installed separately)', value: 'lxml'}
@@ -1211,6 +1234,8 @@ function ConfigFields() {
                             name: null,
                             showOnSearch: true,
                             score: 0,
+                            username: null,
+                            password: null,
                             preselect: true,
                             accessType: "both",
                             search_ids: ["imdbid", "rid", "tvdbid"],