+++ static/js/nzbhydra.js
@@ -3047,6 +3047,26 @@ function ConfigFields() {
                 },
                 {
                     wrapper: 'fieldset',
+                    templateOptions: {label: 'UI'},
+                    fieldGroup: [
+
+                        {
+                            key: 'theme',
+                            type: 'horizontalSelect',
+                            templateOptions: {
+                                type: 'select',
+                                label: 'Theme',
+                                help: 'Reload page after saving',
+                                options: [
+                                    {name: 'Default', value: 'default'},
+                                    {name: 'Dark', value: 'dark'}
+                                ]
+                            }
+                        }
+                    ]
+                },
+                {
+                    wrapper: 'fieldset',
                     templateOptions: {label: 'Security'},
                     fieldGroup: [
 
@@ -3121,6 +3141,7 @@ function ConfigFields() {
 
                     ]
                 },
+                
                 {
                     wrapper: 'fieldset',
                     templateOptions: {label: 'Other'},